[test visibility] Reorg test visibility e2e tests for mocha and jest (#4456)
